+++ frontend/src/pages/AdminPage/tabs/ClubInfoEditTab/components/mobile/LinkEditPage/LinkField.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import FieldClearButtonIcon from '@/assets/images/icons/field_clear_button_icon.svg?react';
 import { colors } from '@/styles/theme/colors';
-import EditField from '../EditField/EditField';
+import EditField from '@/pages/AdminPage/components/editFields/EditField/EditField';
 import * as Styled from './LinkField.styles';
 
 interface LinkFieldProps {
